feat : 버튼 배치 변경
#757
