+++ src/pages/Study/Accordion/StudyAccordionHeader.tsx
@@ -1,29 +1,13 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Chip, Divider, Typography } from '@mui/material';
-import { PeriodicInfo, StudyInfo } from '@api/dto';
-import useCheckAuth from '@hooks/useCheckAuth';
-import ActionButton from '@components/Button/ActionButton';
+import { StudyInfo } from '@api/dto';
 import ServerImg from '@components/Image/ServerImg';
-import StudyModal from '../Modal/StudyModal';
 
 interface StudyAccordionHeaderProps {
   study: StudyInfo;
-  currentPeriod: PeriodicInfo;
 }
 
-const StudyAccordionHeader = ({ study, currentPeriod }: StudyAccordionHeaderProps) => {
-  const [studyModalOpen, setStudyModalOpen] = useState(false);
-
-  const { checkIsMyId } = useCheckAuth();
-
-  const handleStudyEditButtonClick = () => {
-    setStudyModalOpen(true);
-  };
-
-  const handleStudyDeleteButtonClick = () => {
-    // TODO 스터디 제거 API 호출 후 새로고침(기능 구현 후 console 삭제 예정)
-  };
-
+const StudyAccordionHeader = ({ study }: StudyAccordionHeaderProps) => {
   return (
     <div className="flex w-full items-center justify-between px-2">
       <div className="flex items-center space-x-4">
@@ -38,16 +22,6 @@ const StudyAccordionHeader = ({ study, currentPeriod }: StudyAccordionHeaderProp
         <Typography variant="h3" fontWeight="semiBold">
           {study.title}
         </Typography>
-        {checkIsMyId(study.headId) && (
-          <div className="space-x-2">
-            <ActionButton mode="edit" small onClick={handleStudyEditButtonClick}>
-              수정
-            </ActionButton>
-            <ActionButton mode="delete" small onClick={handleStudyDeleteButtonClick}>
-              삭제
-            </ActionButton>
-          </div>
-        )}
       </div>
       <div className="flex items-center space-x-2">
         <Typography>스터디장</Typography>
@@ -57,12 +31,6 @@ const StudyAccordionHeader = ({ study, currentPeriod }: StudyAccordionHeaderProp
           현재 인원 <span className="font-semibold">{study.memberCount}명</span>
         </Typography>
       </div>
-      <StudyModal
-        open={studyModalOpen}
-        setOpen={setStudyModalOpen}
-        selectedStudyInfo={study}
-        currentPeriod={currentPeriod}
-      />
     </div>
   );
 };
